chore(admin): convert ProfilePage and re-use server driven types
