--- packages/core/admin/admin/src/pages/ProfilePage/utils/schema.js
@@ -1,7 +1,0 @@
-import * as yup from 'yup';
-
-import { profileValidation } from '../../SettingsPage/pages/Users/utils/validations/users';
-
-const schema = yup.object().shape(profileValidation);
-
-export default schema;
